# product-api/handlers/product.go
// Package classification Petstore API.
//
// Documentation of API
//
//     Schemes: http
//     Host: localhost
//     BasePath: /
//     Version: 1.0.0
//     Contact: Kamal SELVAM<kselvam.phd@gmail.com>
//
//     Consumes:
//     - application/json
//
//     Produces:
//     - application/json
//
// swagger:meta
package handlers

import (
	"log"
)

type Products struct {
	l *log.Logger
}

func NewProduct(l *log.Logger) *Products {
	return &Products{l}
}

type KeyProduct struct {
}

// GenericError is a generic error message returned by a server
type GenericError struct {
	Message string `json:"message"`
}

// ValidationError is a collection of validation error messages
type ValidationError struct {
	Messages []string `json:"messages"`
}